Tests/LibWeb/Text/input/HTML/png-without-IDAT-should-still-load.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(done => {
        const image = document.createElement("img");
        image.onload = () => {
            println(`loaded, width=${image.width}, height=${image.height}`);
            done();
        };
        image.onerror = () => {
            println("error :^(");
            done();
        };

        image.src = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAD4AAABhCAIAAAABe4UxAAAABElEQVQAAAABnSTXkQAAAABJRU5ErkJggg==";
    });
</script>
