Tests/LibWeb/Text/expected/HTML/png-without-IDAT-should-still-load.txt

loaded, width=62, height=97
